+++ src/prefect/engine/task_runner.py
@@ -29,6 +29,7 @@ from prefect.engine.result_handlers import JSONResultHandler, ResultHandler
 from prefect.engine.runner import ENDRUN, Runner, call_state_handlers
 from prefect.engine.state import (
     Cached,
+    Cancelled,
     Failed,
     Looped,
     Mapped,
@@ -868,6 +869,11 @@ class TaskRunner(Runner):
             result = timeout_handler(
                 self.task.run, timeout=self.task.timeout, **raw_inputs
             )
+
+        except KeyboardInterrupt:
+            self.logger.exception("Interrupt signal raised, cancelling task run.")
+            state = Cancelled(message="Interrupt signal raised, cancelling task run.")
+            return state
 
         # inform user of timeout
         except TimeoutError as exc: